CDDSO-659 add a common set of processors in mip convert and allow the mappings plugins to override them (#483)
* CDDSO-659 Move processors to base plugin

* CDDSO-659 Remove duplicate processors

* CDDSO-659 Update imports

* CDDSO-659 Update test import

* CDDSO-659 Add plugin overrides.
