+++ mip_convert/mip_convert/plugins/hadrem_cp4a/hadrem_cp4a_plugin.py
@@ -10,6 +10,7 @@ from typing import Dict, Any
 import iris.cube
 
 from mip_convert.plugins.base.base_plugin import BaseMappingPlugin
+from mip_convert.plugins.base.data.processors import *
 from mip_convert.plugins.hadrem_cp4a.data.processors import *
